Add a metadata generation for keys removed between two versions
This commit adds a generator for keys that were removed between two
versions. This use the new "deprecation level feature" introduced in
Spring Boot 2.

A deprecated key can still be defined with an error level. In that case
the IDE should still resolve the key, but as an error. This should
improve the developer experience of those who are upgrading and are
wondering what they should change in their configuration.